+++ kotlintest-core/build.gradle
@@ -1,5 +1,4 @@
 dependencies {
-    compile project(':kotlintest-assertions')
     compile 'org.slf4j:slf4j-api:1.7.25'
     compile "org.jetbrains.kotlin:kotlin-reflect"
     compile 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.1.1'
