--- kotlintest-extensions/kotlintest-extensions-junitxml/build.gradle
@@ -1,12 +1,0 @@
-dependencies {
-    compile project(':kotlintest-core')
-    compile project(':kotlintest-assertions')
-    testCompile project(':kotlintest-runner:kotlintest-runner-junit5')
-}
-
-test {
-    useJUnitPlatform()
-    testLogging {
-        events "PASSED", "FAILED", "SKIPPED", "STANDARD_OUT", "STANDARD_ERROR"
-    }
-}
