--- kotlintest-extensions/kotlintest-extensions-system/build.gradle
@@ -1,14 +1,0 @@
-dependencies {
-    compile project(':kotlintest-core')
-    compile project(':kotlintest-assertions')
-    testCompile project(':kotlintest-runner:kotlintest-runner-junit5')
-    testRuntime 'log4j:log4j:1.2.17'
-    testRuntime 'org.slf4j:slf4j-log4j12:1.7.25'
-}
-
-test {
-    useJUnitPlatform()
-    testLogging {
-        events "PASSED", "FAILED", "SKIPPED", "STANDARD_OUT", "STANDARD_ERROR"
-    }
-}
