+++ kotlintest-extensions/src/main/kotlin/io/kotlintest/extensions/system/NoSystemOutListener.kt
