+++ kotlintest-extensions/src/test/kotlin/io/kotlintest/extensions/system/NoSytemOutOrErrTest.kt
