+++ kotlintest-extensions/src/test/kotlin/io/kotlintest/extensions/system/SystemExitTest.kt
