+++ kotlintest-extensions/src/test/resources/log4j2.xml
