+++ kotlintest-runner/kotlintest-runner-jvm/build.gradle
@@ -1,5 +1,6 @@
 dependencies {
     compile project(':kotlintest-core')
+    compile project(":kotlintest-extensions")
     compile 'io.github.classgraph:classgraph:4.6.20'
     compile 'org.slf4j:slf4j-api:1.7.25'
     compile 'io.arrow-kt:arrow-core:0.8.2'
