+++ settings.gradle
@@ -8,9 +8,8 @@ include 'kotlintest-core',
         'kotlintest-runner:kotlintest-runner-junit4',
         'kotlintest-runner:kotlintest-runner-junit5',
         'kotlintest-extensions:kotlintest-extensions-spring',
-        'kotlintest-extensions:kotlintest-extensions-junitxml',
         'kotlintest-extensions:kotlintest-extensions-allure',
-        'kotlintest-extensions:kotlintest-extensions-system',
+        'kotlintest-extensions:kotlintest-extensions',
         'kotlintest-plugin-intellij',
         'kotlintest-tests:kotlintest-tests-core',
         'kotlintest-tests:kotlintest-tests-junit-report',
